src/AIDocumentPipeline/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.20.0
azure-functions-durable==1.2.9
azure-identity==1.17.1
azure-storage-blob==12.22.0
openai==1.40.1
pdf2image==1.17.0
